+++ package.json
@@ -19,7 +19,7 @@
     "g": "yarn codegen",
     "typecheck": "tsc -p . --noEmit",
     "tc": "yarn typecheck",
-    "barrelsby": "barrelsby -d ./src -D -i \"src/(components|hooks|reducers|types)/(?!.*\\.d\\.ts$).*$\" -l below -q",
+    "barrelsby": "barrelsby -d ./src -D -i \"src/(components|contexts|hooks|reducers|types)/(?!.*\\.d\\.ts$).*$\" -l below -q",
     "b": "yarn barrelsby",
     "prepare": "husky install"
   },
@@ -29,31 +29,31 @@
     "@material-ui/lab": "4.0.0-alpha.57",
     "@material-ui/styles": "4.11.3",
     "clsx": "1.1.1",
-    "gatsby": "3.1.1",
+    "gatsby": "3.1.2",
     "gatsby-image": "3.1.0",
     "gatsby-plugin-advanced-sitemap": "1.6.0",
     "gatsby-plugin-csp": "1.1.3",
     "gatsby-plugin-google-analytics": "3.1.0",
-    "gatsby-plugin-image": "1.1.1",
+    "gatsby-plugin-image": "1.1.2",
     "gatsby-plugin-manifest": "3.1.0",
     "gatsby-plugin-offline": "4.1.0",
     "gatsby-plugin-react-i18next": "1.1.0",
     "gatsby-plugin-root-import": "2.0.6",
-    "gatsby-plugin-sharp": "3.1.1",
+    "gatsby-plugin-sharp": "3.1.2",
     "gatsby-plugin-ts-config": "1.1.5",
-    "gatsby-source-contentful": "5.1.1",
+    "gatsby-source-contentful": "5.1.2",
     "gatsby-source-filesystem": "3.1.0",
     "gatsby-theme-material-ui": "1.0.13",
     "gatsby-transformer-inline-svg": "0.1.0",
     "gatsby-transformer-remark": "3.1.0",
     "gatsby-transformer-sharp": "3.1.0",
     "html-react-parser": "1.2.4",
-    "i18next": "19.9.2",
+    "i18next": "20.1.0",
     "js-cookie": "2.2.1",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-helmet": "6.1.0",
-    "react-i18next": "11.8.10"
+    "react-i18next": "11.8.11"
   },
   "devDependencies": {
     "@babel/cli": "7.13.10",
@@ -64,10 +64,10 @@
     "@types/js-cookie": "2.2.6",
     "@types/node": "14.14.35",
     "@types/react": "17.0.3",
-    "@types/react-dom": "17.0.2",
+    "@types/react-dom": "17.0.3",
     "@types/react-helmet": "6.1.0",
-    "@typescript-eslint/eslint-plugin": "4.18.0",
-    "@typescript-eslint/parser": "4.18.0",
+    "@typescript-eslint/eslint-plugin": "4.19.0",
+    "@typescript-eslint/parser": "4.19.0",
     "babel-plugin-i18next-extract": "0.8.2",
     "babel-preset-gatsby": "1.1.0",
     "barrelsby": "2.2.0",
@@ -80,10 +80,10 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.3.1",
     "eslint-plugin-promise": "4.3.1",
-    "eslint-plugin-react": "7.22.0",
+    "eslint-plugin-react": "7.23.1",
     "eslint-plugin-react-hooks": "4.2.0",
     "eslint-plugin-standard": "4.1.0",
-    "husky": "5.1.3",
+    "husky": "5.2.0",
     "prettier": "2.2.1",
     "typescript": "4.2.3"
   },
