+++ src/components/index.ts
@@ -6,6 +6,8 @@ export * from './atoms/SvgAvatar';
 export * from './atoms/SvgIcon';
 export * from './molecules/AvatarCard';
 export * from './molecules/ExpansionPanel';
+export * from './molecules/ExpansionPanelDetail';
+export * from './molecules/ExpansionPanelSummary';
 export * from './molecules/MediaCard';
 export * from './organisms/CertificationList';
 export * from './organisms/ContactList';
